Single page design / hex dump
I haven't done much with the front end here, but the game plan is to do
this single page design style. I also hacked together a quick little
python script the reads in a static image and prints the hex value of
the image to a static file. Shit totally works and this pretty much
puts us at a quarter of the way done.